msm: camera2: Enable frame dump when MICRO bad state is hit.
The MICRO can enter bad state continously if the frame message
command that is issued to MICRO is having wrong configurations.
This could be due to multiple configuration parameters sent by
different modules. To identity the root cause, the frame message
command payload is dumped when MICRO bad state is detected.

This change helps detect occurance of:
10863565 [Stability] Camera preview showed up green with vertical bars rolling

Signed-off-by: Shuzhen Wang <shuzhenw@codeaurora.org>