msm: vidc: fix a race condition in dynamic buffer handling
Fixes a race condition while ftb is queued to the driver
and release reference event for the same has been returned
from the firmware. Without this change, we would end up
having a buffer which is registered in the list but not
queued to firmware. As a result during flush this buffer is
not returned back to client after a flush and results in
video freeze.

b/11021131

Signed-off-by: Arun Menon <avmenon@codeaurora.org>