power: max17048: fix having old soc when the system is resumed
Since it takes some time to restart the battery monitor work from
the system resume, there is a brief period when the battery level
has a value before suspend. To fix this, wait until the battery
level is updated before returning the capacity if the battery
monitor work is not started yet.

Change-Id: Ibbebcd653dd476d36a24c6b89ed8e4c0a00391be
Signed-off-by: choongryeol.lee <choongryeol.lee@lge.com>